pyproject.toml
[project]
name = "qr-code-genrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.1.0",
    "qrcode[pil]>=8.0",
    "streamlit>=1.44.0",
]
